+++ code/ext/ext.driver.auth.privy/src/test.ui/-TestRunner.TESTS.ts
@@ -1,7 +1,7 @@
 export const TESTS = {
   get all() {
     return [
-      import('./-TEST.mjs'),
+      import('./-TEST'),
       import('../evm/Chain/Chain.TEST'),
       import('../evm/Balance/Balance.TEST'),
       import('../http/ExchangeRate/-TEST'),
